[AIP-21] Move contrib.operators.wasb_delete_blob_operator airflow.contrib.operators.wasb_delete_blob_operator
